labextension: Fix error reporting
We were trying to show an error that doesn't exists when code cell is
empty. This is because we were doing a truthy check on
`formattedText.code`, empty string is false so we went ahead to try to
show the error.

Fix this by checking if `formattedText.error` is truthy instead.